+++ test/jsonschema/src/jsonschema_draft202012_tests.cpp
@@ -101,14 +101,14 @@ TEST_CASE("jsonschema draft2020-12 tests")
 {
     SECTION("issues")
     {
-        jsonschema_tests("./jsonschema/issues/draft2020-12/issue-uri.json",
-            jsonschema::evaluation_options{}.default_version(jsonschema::schema_version::draft202012()).
-            require_format_validation(true));
-        //jsonschema_tests("./jsonschema/JSON-Schema-Test-Suite/tests/draft2020-12/optional/format/uri.json",
+        //jsonschema_tests("./jsonschema/issues/draft2020-12/issue-uri.json",
         //    jsonschema::evaluation_options{}.default_version(jsonschema::schema_version::draft202012()).
         //    require_format_validation(true));
+        jsonschema_tests("./jsonschema/JSON-Schema-Test-Suite/tests/draft2020-12/optional/format/uri.json",
+            jsonschema::evaluation_options{}.default_version(jsonschema::schema_version::draft202012()).
+            require_format_validation(true));
     }
-#if 0
+
     SECTION("more_tests")
     {
         // unevaluated-tests.json is from https://github.com/networknt/json-schema-validator/tree/master/src/test/resources/schema/unevaluatedTests
@@ -217,12 +217,11 @@ TEST_CASE("jsonschema draft2020-12 tests")
             jsonschema::evaluation_options{}.default_version(jsonschema::schema_version::draft202012()).
                 require_format_validation(true));
 
-        //jsonschema_tests("./jsonschema/JSON-Schema-Test-Suite/tests/draft2020-12/optional/format/uri.json",
-        //    jsonschema::evaluation_options{}.default_version(jsonschema::schema_version::draft202012()).
-        //    require_format_validation(true));
+        jsonschema_tests("./jsonschema/JSON-Schema-Test-Suite/tests/draft2020-12/optional/format/uri.json",
+            jsonschema::evaluation_options{}.default_version(jsonschema::schema_version::draft202012()).
+            require_format_validation(true));
         //jsonschema_tests("./jsonschema/JSON-Schema-Test-Suite/tests/draft2020-12/optional/format/uri-reference.json");
         //jsonschema_tests("./jsonschema/JSON-Schema-Test-Suite/tests/draft2020-12/optional/format/uri-template.json");
         jsonschema_tests("./jsonschema/JSON-Schema-Test-Suite/tests/draft2020-12/content.json");
     }
-#endif
 }