+++ CHANGELOG.md
@@ -52,6 +52,8 @@ json j{jsoncons::null_arg};
 
 - Custom jmespath functions are now supported thanks to PR #560
 
+- jsonschema now understands the uri format
+
 0.178.0
 -------